requirements.txt
contourpy==1.3.2
cycler==0.12.1
einops==0.8.2
fonttools==4.62.0
fsspec==2026.2.0
huggingface_hub==1.6.0
kiwisolver==1.5.0
lmdb==1.5.1
matplotlib==3.10.8
mkl-service==2.4.0
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.0
Pillow==9.0.1
protobuf==3.12.4
pyparsing==3.3.2
python-dateutil==2.8.1
PyYAML==6.0.2
safetensors==0.7.0
scipy==1.15.3
six==1.16.0
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
timm==1.0.25
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
zipp==1.0.0